+++ examples/10_tracking_diagnostics/pet_groups.py
@@ -1,6 +1,6 @@
 """
 Groups distribution
-==================
+===================
 
 """
 import py_eddy_tracker_sample